forgot to commit and push

+++ FTLGame/offsets.h
@@ -1,8 +1,9 @@
 #pragma once
+#include <Windows.h>
 namespace offsets
 {
-	int inDanger = 0x4CA820;
-	int saveClass = 0x4C448C;
-	int inSector = 0x4C4C00;
-	int isAE_Enabled = 0x4C4C10;
+	uintptr_t inDanger = 0x4CA820;
+	uintptr_t saveClass = 0x4C448C;
+	uintptr_t inSector = 0x4C4C00;
+	uintptr_t isAE_Enabled = 0x4C4C10;
 }
